requirements.txt
beautifulsoup4
pyppeteer
flask[async]
python-dotenv
groq
boto3
datetime
requests
Pillow
sentencepiece
transformers
socksio
python-socks
supabase